Don't truncate resource usage
Convert the resource usage to `int64` will usually truncate
the fracion part, this is misleading if the actual value is
less than 1.

+++ staging/src/k8s.io/kubectl/pkg/metricsutil/metrics_printer.go
@@ -295,7 +295,7 @@ func printAllResourceUsages(out io.Writer, metrics *ResourceMetricsInfo) {
 		fmt.Fprint(out, "\t")
 		if available, found := metrics.Available[res]; found {
 			fraction := float64(quantity.MilliValue()) / float64(available.MilliValue()) * 100
-			fmt.Fprintf(out, "%d%%\t", int64(fraction))
+			fmt.Fprintf(out, "%.2f%%\t", fraction)
 		}
 	}
 }
